IMS-VT: Use correct view size when device is rotated.
When peer_resolution received is portrait mode and the device
is in landscape, the far-end video view will be re sized to
fit the peer aspect ratio. If we rotate the device back to
portrait, the view size needs to be re calculated based on
the previously received peer_resolution.

Change-Id: I47b84fd823c6110dca87bcc372be1aaeb9888efa
CRs-Fixed: 887207
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>