IMS-VT: Set preview surface only if camera is opened
Unable to open camera once user went to background and foreground
if PARAM_READY is received when activity is in background due to
preview surface is set as part of PARAM_READY.

If camera is not opened don't set the preview surface.
Flow of setting preview surface will be openCamera->
RequestCameraCapablities->surface set.

Change-Id: I5f5a37ec2f97a14f449dc757f8806b24a9144396
CRs-Fixed: 939557
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>