IMS-VT: Fix for auto fullscreen issue during UI rotation.
Consider case where VT call UI is in fullscreen mode and the
UI orientation is changed. When InCall activity is re created
due to rotation, the fullscreen flag is still true, but actual
UI will be created in normal mode. Due to wrong fullscreen flag
value, auto enter fullscreen thread is not scheduled.
To fix this clear fullscreen flag in InCallActivity.onCreate().

Change-Id: I18ec2b4442eae3b9ec01aeacc2b38f91f08f21c9
CRs-Fixed: 979860
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>